Refactor `MerchantEmailNotifications`
